feat: some refactoring applied
