+++ src/Application/Services/Common/CalendarService.cs
@@ -22,7 +22,7 @@ namespace Application.Services.Common;
 /// <param name="dateTimeService">The date time service instance to use.</param>
 /// <param name="loggerService">The logger service instance to use.</param>
 /// <param name="mapper">The auto mapper instance to use.</param>
-internal sealed class CalendarDayService(IDateTimeService dateTimeService, ILoggerService<CalendarDayService> loggerService, IMapper mapper) : ICalendarDayService
+internal sealed class CalendarService(IDateTimeService dateTimeService, ILoggerService<CalendarService> loggerService, IMapper mapper) : ICalendarService
 {
 	private readonly IQueryable<DateTime> _dateTimes = GetPossibleDates();
