+++ src/Presentation/Controllers/CalendarController.cs
@@ -3,7 +3,7 @@
 using Application.Contracts.Responses.Common;
 using Application.Features.Requests;
 using Application.Features.Responses;
-using Application.Interfaces.Application;
+using Application.Interfaces.Application.Common;
 
 using Asp.Versioning;
 
@@ -23,89 +23,58 @@ namespace Presentation.Controllers;
 /// <remarks>
 /// Inherits from <see cref="ApiControllerBase"/>.
 /// </remarks>
-/// <param name="calendarDayService">The calendar day service.</param>
+/// <param name="calendarService">The calendar service instance to use.</param>
 [Route(Endpoints.Calendar.BaseUri)]
 [ApiVersion(Versioning.CurrentVersion)]
-public sealed class CalendarController(ICalendarService calendarDayService) : ApiControllerBase
+public sealed class CalendarController(ICalendarService calendarService) : ApiControllerBase
 {
-	private readonly ICalendarService _calendarDayService = calendarDayService;
+	private readonly ICalendarService _calendarService = calendarService;
 
 	/// <summary>
 	/// Returns calendar entries as a paged list, filtered by the parameters.
 	/// </summary>
-	/// <param name="parameters">The calendar query parameters.</param>
-	/// <param name="cancellationToken">The cancellation token to cancel the request.</param>
-	/// <response code="200">If the result is returned.</response>
-	/// <response code="404">If the result is empty.</response>
-	/// <response code="500">If something went wrong.</response>
+	/// <param name="parameters">The calendar query parameters to use.</param>
+	/// <response code="200">If the response was successfully returned.</response>
+	/// <response code="500">Something internal went terribly wrong.</response>
 	[HttpGet(Endpoints.Calendar.GetPagedByParameters)]
 	[ProducesResponseType(typeof(IPagedList<CalendarResponse>), StatusCodes.Status200OK)]
 	[ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status404NotFound)]
 	[ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
-	public async Task<IActionResult> GetPagedByParameters([FromQuery] CalendarParameters parameters, CancellationToken cancellationToken)
+	public IActionResult GetPagedByParameters([FromQuery] CalendarParameters parameters)
 	{
-		ErrorOr<IPagedList<CalendarResponse>> result =
-			await _calendarDayService.Get(parameters, false, cancellationToken);
-
-		return Get(result, result.Value?.MetaData);
+		ErrorOr<IPagedList<CalendarResponse>> result = _calendarService.GetPagedByParameters(parameters);
+		return Get(result, result.Value.MetaData);
 	}
 
 	/// <summary>
 	/// Returns the calendar entry by its date.
 	/// </summary>
 	/// <param name="date">The date of the calendar entry.</param>
-	/// <param name="cancellationToken">The cancellation token to cancel the request.</param>
-	/// <response code="200">If the result is returned.</response>
-	/// <response code="404">If the result is empty.</response>
-	/// <response code="500">If something went wrong.</response>
+	/// <response code="200">If the response was successfully returned.</response>
+	/// <response code="404">The requested resource could not be found.</response>
+	/// <response code="500">Something internal went terribly wrong.</response>
 	[HttpGet(Endpoints.Calendar.GetByDate)]
 	[ProducesResponseType(typeof(CalendarResponse), StatusCodes.Status200OK)]
 	[ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status404NotFound)]
 	[ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
-	public async Task<IActionResult> GetByDate([DataType(DataType.Date)] DateTime date, CancellationToken cancellationToken)
+	public IActionResult GetByDate([DataType(DataType.Date)] DateTime date)
 	{
-		ErrorOr<CalendarResponse> result =
-			await _calendarDayService.Get(date, false, cancellationToken);
-
-		return Get(result);
-	}
-
-	/// <summary>
-	/// Returns the calendar entry by its identifier.
-	/// </summary>
-	/// <param name="id">The identifier of the calendar entry.</param>
-	/// <param name="cancellationToken">The cancellation token to cancel the request.</param>
-	/// <response code="200">If the result is returned.</response>
-	/// <response code="404">If the result is empty.</response>
-	/// <response code="500">If something went wrong.</response>
-	[HttpGet(Endpoints.Calendar.GetById)]
-	[ProducesResponseType(typeof(CalendarResponse), StatusCodes.Status200OK)]
-	[ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status404NotFound)]
-	[ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
-	public async Task<IActionResult> GetById(Guid id, CancellationToken cancellationToken)
-	{
-		ErrorOr<CalendarResponse> result =
-			await _calendarDayService.Get(id, false, cancellationToken);
-
+		ErrorOr<CalendarResponse> result = _calendarService.GetByDate(date);
 		return Get(result);
 	}
 
 	/// <summary>
 	/// Returns the current calendar entry.
 	/// </summary>
-	/// <param name="cancellationToken">The cancellation token to cancel the request.</param>
-	/// <response code="200">If the result is returned.</response>
-	/// <response code="404">If the result is empty.</response>
-	/// <response code="500">If something went wrong.</response>
+	/// <response code="200">If the response was successfully returned.</response>
+	/// <response code="500">Something internal went terribly wrong.</response>
 	[HttpGet(Endpoints.Calendar.GetCurrent)]
 	[ProducesResponseType(typeof(CalendarResponse), StatusCodes.Status200OK)]
 	[ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status404NotFound)]
 	[ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
-	public async Task<IActionResult> GetCurrent(CancellationToken cancellationToken)
+	public IActionResult GetCurrent()
 	{
-		ErrorOr<CalendarResponse> result =
-			await _calendarDayService.Get(false, cancellationToken);
-
+		ErrorOr<CalendarResponse> result = _calendarService.GetCurrent();
 		return Get(result);
 	}
 }
