+++ tests/ApplicationTests/Services/Common/CalendarServiceTests.cs
@@ -18,16 +18,16 @@ namespace ApplicationTests.Services.Common;
 
 [TestClass]
 [SuppressMessage("Style", "IDE0058", Justification = "Not relevant here, unit tests.")]
-public sealed class CalendarDayServiceTests : ApplicationTestBase
+public sealed class CalendarServiceTests : ApplicationTestBase
 {
 	private readonly IMapper _mapper = GetService<IMapper>();
 	private Mock<IDateTimeService> _dateTimeServiceMock = default!;
-	private Mock<ILoggerService<CalendarDayService>> _loggerServiceMock = default!;
+	private Mock<ILoggerService<CalendarService>> _loggerServiceMock = default!;
 
 	[TestMethod, TestCategory("Methods")]
 	public void GetByDateShouldReturnResponseWhenSuccessfully()
 	{
-		CalendarDayService sut = CreateMockedInstance();
+		CalendarService sut = CreateMockedInstance();
 		DateTime date = DateTime.Today;
 
 		ErrorOr<CalendarResponse> result = sut.GetByDate(date);
@@ -40,7 +40,7 @@ public sealed class CalendarDayServiceTests : ApplicationTestBase
 	[TestMethod, TestCategory("Methods")]
 	public void GetByDateShouldReturnFailedResponseWhenExcpetionGetThrown()
 	{
-		CalendarDayService sut = CreateMockedInstance();
+		CalendarService sut = CreateMockedInstance();
 		DateTime date = DateTime.MinValue;
 
 		ErrorOr<CalendarResponse> result = sut.GetByDate(date);
@@ -53,7 +53,7 @@ public sealed class CalendarDayServiceTests : ApplicationTestBase
 	[TestMethod, TestCategory("Methods")]
 	public void GetPagedByParametersShouldReturnResponseWhenSuccessfully()
 	{
-		CalendarDayService sut = CreateMockedInstance();
+		CalendarService sut = CreateMockedInstance();
 		CalendarParameters parameters = new() { Year = 2020, PageSize = 100 };
 
 		ErrorOr<IPagedList<CalendarResponse>> result = sut.GetPagedByParameters(parameters);
@@ -69,7 +69,7 @@ public sealed class CalendarDayServiceTests : ApplicationTestBase
 	[TestMethod, TestCategory("Methods")]
 	public void GetPagedByParametersShouldReturnFailedResponseWhenExcpetionGetThrown()
 	{
-		CalendarDayService sut = CreateMockedInstance();
+		CalendarService sut = CreateMockedInstance();
 
 		ErrorOr<IPagedList<CalendarResponse>> result = sut.GetPagedByParameters(null!);
 
@@ -82,7 +82,7 @@ public sealed class CalendarDayServiceTests : ApplicationTestBase
 	public void GetCurrentShouldReturnResponseWhenSuccessfully()
 	{
 		DateTime today = DateTime.Today;
-		CalendarDayService sut = CreateMockedInstance();
+		CalendarService sut = CreateMockedInstance();
 		_dateTimeServiceMock.Setup(x => x.Today).Returns(today);
 
 		ErrorOr<CalendarResponse> result = sut.GetCurrent();
@@ -95,7 +95,7 @@ public sealed class CalendarDayServiceTests : ApplicationTestBase
 	[TestMethod, TestCategory("Methods")]
 	public void GetCurrentShouldReturnFailedResponseWhenExcpetionGetThrown()
 	{
-		CalendarDayService sut = CreateMockedInstance();
+		CalendarService sut = CreateMockedInstance();
 		_dateTimeServiceMock.Setup(x => x.Today).Returns(DateTime.MinValue);
 
 		ErrorOr<CalendarResponse> result = sut.GetCurrent();
@@ -105,7 +105,7 @@ public sealed class CalendarDayServiceTests : ApplicationTestBase
 		result.Errors.First().Should().Be(CalendarServiceErrors.GetCurrentDateFailed);
 	}
 
-	private CalendarDayService CreateMockedInstance()
+	private CalendarService CreateMockedInstance()
 	{
 		_dateTimeServiceMock = new();
 		_loggerServiceMock = new();
